alicorp-iagen-mvp/requirements.txt
streamlit>=1.34
python-dotenv>=1.0
pandas>=2.2
numpy>=1.26

# Gemini
google-genai>=0.3.0
google-cloud-aiplatform>=1.70
pillow>=10.3

